centralized MultiHex version number and about dialog.
